linear quant code done; UNDO: fix accuracy and implement outer interface

+++ csrc/normalization.cpp
@@ -124,8 +124,9 @@ at::Tensor i_residual_layernorm_ (
   for (auto i = 0; i < batch; ++i) {
     auto* psrc1 = reinterpret_cast<int8_t (*)[reduce_l]>(src1);
     auto* psrc2 = reinterpret_cast<int8_t (*)[reduce_l]>(src2);
-    auto* pw = reinterpret_cast<float *>(w);
-    auto* pb = reinterpret_cast<float *>(b);
+    // TODO: add switch float or _Float16
+    auto* pw = reinterpret_cast<_Float16 *>(w);
+    auto* pb = reinterpret_cast<_Float16 *>(b);
     auto* pout = reinterpret_cast<int8_t (*)[reduce_l]>(out);
 
     i_residual_layernorm_tpp<32>::ref_fp16(
